sped up editing and rerendering (major change)
modified the code to only render the new elements when they are added.
this involved creating a new variable called brackets.textbookContainer
used to identify it. all rendering is the done in the "add" methods in
textbook.js. tried to do this whenever possible, but i may have missed
some bits. its much faster on my computer now.